+++ theme/types/index.d.ts
@@ -9,7 +9,7 @@ export interface ThemeConfig {
   colors?: {
     /**
      * primary color
-     * @default '#e67474'
+     * @default '#0078E7'
      */
     primary: string
   }